requirements.txt
absl-py==2.1.0
aim==3.18.1
aim-ui==3.18.1
aimrecords==0.0.7
aimrocks==0.4.0
aiofiles==23.2.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
astroid @ file:///tmp/build/80754af9/astroid_1639030679373/work
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
base58==2.0.1
brotlipy==0.7.0
cachetools==5.3.2
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
contourpy==1.1.1
cryptography @ file:///tmp/build/80754af9/cryptography_1652083738073/work
cycler==0.12.1
debugpy @ file:///tmp/build/80754af9/debugpy_1637091796427/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
docker-pycreds==0.4.0
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649926445639/work
exceptiongroup==1.2.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastapi==0.109.2
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.28.0
google-auth-oauthlib==1.0.0
greenlet==3.0.3
grpcio==1.60.1
h11==0.14.0
huggingface-hub==0.20.3
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.34.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
iopath==0.1.10
ipykernel @ file:///tmp/build/80754af9/ipykernel_1646982648551/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///opt/conda/conda-bld/ipython_1651600145335/work
isort @ file:///tmp/build/80754af9/isort_1628603791788/work
jedi @ file:///tmp/build/80754af9/jedi_1644315233700/work
Jinja2==3.1.3
joblib @ file:///croot/joblib_1685113087166/work
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1650622202839/work
jupyter-core @ file:///opt/conda/conda-bld/jupyter_core_1651671229925/work
kiwisolver==1.4.5
lazy-object-proxy @ file:///tmp/build/80754af9/lazy-object-proxy_1616526917483/work
lazy_loader==0.3
Mako==1.3.2
Markdown==3.5.2
MarkupSafe==2.1.5
-e git+https://github.com/microsoft/GLIP.git@9dda9558c1ef59bb6cdc8e896e2bcab775a68ff0#egg=maskrcnn_benchmark
matplotlib==3.7.5
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mccabe @ file:///opt/conda/conda-bld/mccabe_1644221741721/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1649847908682/work
networkx==3.1
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1652801679809/work
nvidia-ml-py==12.535.133
nvitop==1.3.2
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==23.2
pandas==2.0.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
pillow==10.2.0
platformdirs @ file:///tmp/build/80754af9/platformdirs_1638968569495/work
portalocker==2.8.2
progress==1.6
progressbar==2.5
progressbar2==4.4.2
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==4.25.3
psutil==5.9.8
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.6.1
pydantic_core==2.16.2
Pygments @ file:///croot/pygments_1684279966437/work
pylint @ file:///tmp/build/80754af9/pylint_1647941749960/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.1.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-utils==3.8.2
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq @ file:///tmp/build/80754af9/pyzmq_1638436375034/work
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.1
RestrictedPython==7.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sentry-sdk==1.40.5
setproctitle==1.3.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.27
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.36.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tifffile==2023.7.10
timm==0.6.12
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tornado @ file:///tmp/build/80754af9/tornado_1606942300299/work
tqdm==4.66.2
traitlets @ file:///croot/traitlets_1671143879854/work
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.18
uvicorn==0.27.1
wandb==0.16.3
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
Werkzeug==3.0.1
wrapt @ file:///tmp/build/80754af9/wrapt_1638433239696/work
yacs==0.1.8
zipp==3.17.0
